elastic/etl/requirements.txt
elasticsearch==7.9.1
flake8==4.0
psycopg2-binary==2.9.3
pydantic==1.9.0